pyproject.toml
[tool.poetry]
name = "latex-template-engine"
version = "0.1.0"
description = "A LaTeX template engine with Jinja2 templating and editor integrations"
authors = ["David Dunnock <your.email@example.com>"]
readme = "README.md"
packages = [{include = "latex_template_engine", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
jinja2 = "^3.1.2"
pyyaml = "^6.0"
click = "^8.1.7"
rich = "^13.7.0"
watchdog = "^3.0.0"
pydantic = "^2.5.0"
pathlib2 = "^2.3.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.6.0"

[tool.poetry.scripts]
latex-engine = "latex_template_engine.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
